refactor: OverlapOp + FallbackOp move to operations/overlap.py (AIC-535 / Phase 4.3)
Phase 4 — relocates the two composite ops out of
``operations/_legacy.py`` into a new ``operations/overlap.py``. Both
ops wrap inner ``Operation`` instances and own no CSV data of their
own, so this is pure relocation:

- No ``_data_cache`` / ``load_data`` / ``clear_cache`` — the
  ``Operation`` base class provides empty defaults that suffice.
- No new eager ``OpClass.load_data(self)`` call in
  ``PerfDatabase.__init__``.
- No ``PerfDatabase.query_*`` delegation method — these ops bypass the
  database query layer entirely; they invoke their inner ops'
  ``query()`` directly.

Behavior preserved verbatim:
- ``FallbackOp`` still mutates ``database._default_database_mode``
  during the primary attempt (forces SILICON inside HYBRID) and
  restores via ``finally``; suppresses ERROR-level perf_database logs
  during the primary attempt.
- ``OverlapOp`` still computes ``latency = max(sum_a, sum_b)`` and
  ``energy = sum_a + sum_b`` with source merging.

Logger: same late-binding ``_LoggerProxy`` over ``_legacy.logger`` as
``operations/moe.py`` so existing ``mock.patch`` targets keep working.

After this commit, ``operations/_legacy.py`` contains only:
CustomAllReduce, NCCL, P2P, ContextMLA, GenerationMLA, MLABmm,
MLAModule, WideEPContextMLA, WideEPGenerationMLA, ContextDSAModule,
GenerationDSAModule, Mamba2Kernel, GDNKernel, Mamba2,
_BaseDeepSeekV4AttentionModule, ContextDeepSeekV4AttentionModule,
GenerationDeepSeekV4AttentionModule, DeepSeekV4MHCModule.
All 18 are migrated on other open Phase 2/3 PRs (#1132 / #1136 /
merges.
Verification:
- 670 unit tests pass + 1 skipped (same as baseline)
- 38 regression snapshots bit-identical + 6 skipped (same as baseline)
- ruff check + format clean
